requirements.txt
google-generativeai>=0.3.0
jinja2>=3.1.2
gspread>=5.12.0
python-dotenv>=1.0.0
requests>=2.31.0
oauth2client>=4.1.3
